binary_lt/requirements.txt
absl-py
addict
attrs
cachetools
certifi
charset-normalizer
click
cycler
Deprecated
filelock
fonttools
google-auth
google-auth-oauthlib
grpcio
idna
importlib-metadata
importlib-resources
joblib
jsonschema
kiwisolver
Markdown
matplotlib
mmcv==1.3.13
msgpack
numpy
oauthlib
opencv-python
packaging
pandas
Pillow
protobuf
pyasn1
pyasn1-modules
pyparsing
pyrsistent
python-dateutil
pytz
PyYAML
ray
redis
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
tabulate
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorboardX
threadpoolctl
timm
torch
torchaudio
torchvision
tqdm
typing_extensions
urllib3
Werkzeug
wrapt
yapf
zipp
einops
torchinfo
gdown